UOELSA-1295: Koejakson keskustelut vastuuhenkilölle
