+++ src/main/kotlin/fi/elsapalvelu/elsa/repository/KoejaksonKehittamistoimenpiteetRepository.kt
@@ -24,9 +24,9 @@ interface KoejaksonKehittamistoimenpiteetRepository :
 
     @Query(
         "select kt from KoejaksonKehittamistoimenpiteet kt " +
-            "where kt.id = :id and kt.lahikouluttajaHyvaksynyt = true and kt.lahiesimiesHyvaksynyt = true " +
+            "where kt.id = :id and (kt.lahikouluttaja.user.id = :vastuuhenkiloUserId or kt.lahiesimies.user.id = :vastuuhenkiloUserId or kt.lahikouluttajaHyvaksynyt = true and kt.lahiesimiesHyvaksynyt = true " +
             "and kt.opintooikeus.id in (select va.opintooikeus.id from KoejaksonVastuuhenkilonArvio va " +
-            "where va.vastuuhenkilo.user.id = :vastuuhenkiloUserId)"
+            "where va.vastuuhenkilo.user.id = :vastuuhenkiloUserId))"
     )
     fun findOneByIdHyvaksyttyAndBelongsToVastuuhenkilo(
         id: Long,
