+++ src/main/kotlin/fi/elsapalvelu/elsa/repository/KoejaksonLoppukeskusteluRepository.kt
@@ -25,9 +25,9 @@ interface KoejaksonLoppukeskusteluRepository : JpaRepository<KoejaksonLoppukesku
 
     @Query(
         "select lk from KoejaksonLoppukeskustelu lk " +
-            "where lk.id = :id and lk.lahikouluttajaHyvaksynyt = true and lk.lahiesimiesHyvaksynyt = true " +
+            "where lk.id = :id and (lk.lahikouluttaja.user.id = :vastuuhenkiloUserId or lk.lahiesimies.user.id = :vastuuhenkiloUserId or lk.lahikouluttajaHyvaksynyt = true and lk.lahiesimiesHyvaksynyt = true " +
             "and lk.opintooikeus.id in (select va.opintooikeus.id from KoejaksonVastuuhenkilonArvio va " +
-            "where va.vastuuhenkilo.user.id = :vastuuhenkiloUserId)"
+            "where va.vastuuhenkilo.user.id = :vastuuhenkiloUserId))"
     )
     fun findOneByIdHyvaksyttyAndBelongsToVastuuhenkilo(
         id: Long,
