+++ src/main/kotlin/fi/elsapalvelu/elsa/web/rest/kouluttaja/KouluttajaKoejaksoResource.kt
@@ -3,18 +3,13 @@ package fi.elsapalvelu.elsa.web.rest.kouluttaja
 import fi.elsapalvelu.elsa.service.*
 import fi.elsapalvelu.elsa.service.dto.*
 import fi.elsapalvelu.elsa.web.rest.errors.BadRequestAlertException
-import org.springframework.beans.factory.annotation.Value
+import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import tech.jhipster.web.util.ResponseUtil
 import java.security.Principal
-import jakarta.validation.Valid
 
 private const val ENTITY_KOEJAKSON_SOPIMUS = "koejakson_koulutussopimus"
-private const val ENTITY_KOEJAKSON_ALOITUSKESKUSTELU = "koejakson_aloituskeskustelu"
-private const val ENTITY_KOEJAKSON_VALIARVIOINTI = "koejakson_valiarviointi"
-private const val ENTITY_KOEJAKSON_KEHITTAMISTOIMENPITEET = "koejakson_kehittamistoimenpiteet"
-private const val ENTITY_KOEJAKSON_LOPPUKESKUSTELU = "koejakson_loppukeskustelu"
 
 @RestController
 @RequestMapping("/api/kouluttaja")
@@ -27,9 +22,6 @@ class KouluttajaKoejaksoResource(
     private val koejaksonLoppukeskusteluService: KoejaksonLoppukeskusteluService,
     private val koejaksonVaiheetService: KoejaksonVaiheetService
 ) {
-
-    @Value("\${jhipster.clientApp.name}")
-    private var applicationName: String? = null
 
     @GetMapping("/koejaksot")
     fun getKoejaksot(principal: Principal?): ResponseEntity<List<KoejaksonVaiheDTO>> {
@@ -54,7 +46,13 @@ class KouluttajaKoejaksoResource(
         @Valid @RequestBody koulutussopimusDTO: KoejaksonKoulutussopimusDTO,
         principal: Principal?
     ): ResponseEntity<KoejaksonKoulutussopimusDTO> {
-        validateId(koulutussopimusDTO.id, ENTITY_KOEJAKSON_SOPIMUS)
+        if (koulutussopimusDTO.id == null) {
+            throw BadRequestAlertException(
+                "Virheellinen id",
+                ENTITY_KOEJAKSON_SOPIMUS,
+                "idnull"
+            )
+        }
 
         if (koulutussopimusDTO.vastuuhenkilo?.sopimusHyvaksytty == true
             || koulutussopimusDTO.vastuuhenkilo?.kuittausaika != null
@@ -100,62 +98,6 @@ class KouluttajaKoejaksoResource(
         return ResponseUtil.wrapOrNotFound(aloituskeskusteluDTO)
     }
 
-    @PutMapping("/koejakso/aloituskeskustelu")
-    fun updateAloituskeskustelu(
-        @Valid @RequestBody aloituskeskusteluDTO: KoejaksonAloituskeskusteluDTO,
-        principal: Principal?
-    ): ResponseEntity<KoejaksonAloituskeskusteluDTO> {
-        validateId(aloituskeskusteluDTO.id, ENTITY_KOEJAKSON_ALOITUSKESKUSTELU)
-        val user = userService.getAuthenticatedUser(principal)
-
-        var aloituskeskustelu =
-            koejaksonAloituskeskusteluService.findOneByIdAndLahikouluttajaUserId(
-                aloituskeskusteluDTO.id!!,
-                user.id!!
-            )
-
-        if (!aloituskeskustelu.isPresent) {
-            aloituskeskustelu =
-                koejaksonAloituskeskusteluService.findOneByIdAndLahiesimiesUserId(
-                    aloituskeskusteluDTO.id!!,
-                    user.id!!
-                )
-
-            if (!aloituskeskustelu.isPresent) {
-                throw BadRequestAlertException(
-                    "Koejakson aloituskeskustelua ei löydy.",
-                    ENTITY_KOEJAKSON_ALOITUSKESKUSTELU,
-                    "dataillegal.koejakson-aloituskeskustelua-ei-loydy"
-                )
-            }
-
-            if (aloituskeskustelu.get().lahikouluttaja?.sopimusHyvaksytty != true) {
-                throw BadRequestAlertException(
-                    "Esimies ei saa muokata aloituskeskustelua, " +
-                        "jos kouluttaja ei ole allekirjoittanut sitä",
-                    ENTITY_KOEJAKSON_ALOITUSKESKUSTELU,
-                    "dataillegal.erimies-ei-saa-muokata-aloituskeskustelua-jos-kouluttaja-ei-ole-allekirjoittanut-sita"
-                )
-            }
-        }
-
-        if (aloituskeskustelu.get().lahetetty != true) {
-            throw BadRequestAlertException(
-                "Arviointia ei saa muokata, jos erikoistuva ei ole lähettänyt pyyntöä.",
-                ENTITY_KOEJAKSON_ALOITUSKESKUSTELU,
-                "dataillegal.arviointia-ei-saa-muokata-jos-erikoistuva-ei-ole-lahettanyt-pyyntoa"
-            )
-        }
-
-        validateArviointi(
-            aloituskeskustelu.get().lahiesimies?.sopimusHyvaksytty,
-            ENTITY_KOEJAKSON_ALOITUSKESKUSTELU
-        )
-
-        val result = koejaksonAloituskeskusteluService.update(aloituskeskusteluDTO, user.id!!)
-        return ResponseEntity.ok(result)
-    }
-
     @GetMapping("/koejakso/valiarviointi/{id}")
     fun getValiarviointi(
         @PathVariable id: Long,
@@ -170,54 +112,6 @@ class KouluttajaKoejaksoResource(
                 koejaksonValiarviointiService.findOneByIdAndLahiesimiesUserId(id, user.id!!)
         }
         return ResponseUtil.wrapOrNotFound(valiarviointiDTO)
-    }
-
-    @PutMapping("/koejakso/valiarviointi")
-    fun updateValiarviointi(
-        @Valid @RequestBody valiarviointiDTO: KoejaksonValiarviointiDTO,
-        principal: Principal?
-    ): ResponseEntity<KoejaksonValiarviointiDTO> {
-        validateId(valiarviointiDTO.id, ENTITY_KOEJAKSON_LOPPUKESKUSTELU)
-        val user = userService.getAuthenticatedUser(principal)
-
-        var valiarviointi =
-            koejaksonValiarviointiService.findOneByIdAndLahikouluttajaUserId(
-                valiarviointiDTO.id!!,
-                user.id!!
-            )
-
-        if (!valiarviointi.isPresent) {
-            valiarviointi =
-                koejaksonValiarviointiService.findOneByIdAndLahiesimiesUserId(
-                    valiarviointiDTO.id!!,
-                    user.id!!
-                )
-
-            if (!valiarviointi.isPresent) {
-                throw BadRequestAlertException(
-                    "Koejakson väliarviointia ei löydy.",
-                    ENTITY_KOEJAKSON_VALIARVIOINTI,
-                    "dataillegal.koejakson-valiarviointia-ei-loydy"
-                )
-            }
-
-            if (valiarviointi.get().lahikouluttaja?.sopimusHyvaksytty != true) {
-                throw BadRequestAlertException(
-                    "Esimies ei saa muokata väliarviointia, " +
-                        "jos kouluttaja ei ole allekirjoittanut sitä",
-                    ENTITY_KOEJAKSON_VALIARVIOINTI,
-                    "dataillegal.esimies-ei-saa-muoktata-valiarviointia-jos-kouluttaja-ei-ole-allekirjoittanut-sita"
-                )
-            }
-        }
-
-        validateArviointi(
-            valiarviointi.get().lahiesimies?.sopimusHyvaksytty,
-            ENTITY_KOEJAKSON_VALIARVIOINTI
-        )
-
-        val result = koejaksonValiarviointiService.update(valiarviointiDTO, user.id!!)
-        return ResponseEntity.ok(result)
     }
 
     @GetMapping("/koejakso/kehittamistoimenpiteet/{id}")
@@ -239,55 +133,6 @@ class KouluttajaKoejaksoResource(
         return ResponseUtil.wrapOrNotFound(kehittamistoimenpiteetDTO)
     }
 
-    @PutMapping("/koejakso/kehittamistoimenpiteet")
-    fun updateKehittamistoimenpiteet(
-        @Valid @RequestBody kehittamistoimenpiteetDTO: KoejaksonKehittamistoimenpiteetDTO,
-        principal: Principal?
-    ): ResponseEntity<KoejaksonKehittamistoimenpiteetDTO> {
-        validateId(kehittamistoimenpiteetDTO.id, ENTITY_KOEJAKSON_KEHITTAMISTOIMENPITEET)
-        val user = userService.getAuthenticatedUser(principal)
-
-        var kehittamistoimenpiteet =
-            koejaksonKehittamistoimenpiteetService.findOneByIdAndLahikouluttajaUserId(
-                kehittamistoimenpiteetDTO.id!!,
-                user.id!!
-            )
-
-        if (!kehittamistoimenpiteet.isPresent) {
-            kehittamistoimenpiteet =
-                koejaksonKehittamistoimenpiteetService.findOneByIdAndLahiesimiesUserId(
-                    kehittamistoimenpiteetDTO.id!!,
-                    user.id!!
-                )
-
-            if (!kehittamistoimenpiteet.isPresent) {
-                throw BadRequestAlertException(
-                    "Koejakson kehittämistoimenpiteitä ei löydy.",
-                    ENTITY_KOEJAKSON_KEHITTAMISTOIMENPITEET,
-                    "dataillegal.koejakson-kehittamistoimenpiteita-ei-loydy"
-                )
-            }
-
-            if (kehittamistoimenpiteet.get().lahikouluttaja?.sopimusHyvaksytty != true) {
-                throw BadRequestAlertException(
-                    "Esimies ei saa muokata kehittämistoimenpiteitä, " +
-                        "jos kouluttaja ei ole allekirjoittanut niitä",
-                    ENTITY_KOEJAKSON_KEHITTAMISTOIMENPITEET,
-                    "dataillegal.esimies-ei-saa-muokata-kehittamistoimenpiteita-jos-kouluttaja-ei-ole-allekirjottanut-niita"
-                )
-            }
-        }
-
-        validateArviointi(
-            kehittamistoimenpiteet.get().lahiesimies?.sopimusHyvaksytty,
-            ENTITY_KOEJAKSON_KEHITTAMISTOIMENPITEET
-        )
-
-        val result =
-            koejaksonKehittamistoimenpiteetService.update(kehittamistoimenpiteetDTO, user.id!!)
-        return ResponseEntity.ok(result)
-    }
-
     @GetMapping("/koejakso/loppukeskustelu/{id}")
     fun getLoppukeskustelu(
         @PathVariable id: Long,
@@ -305,76 +150,5 @@ class KouluttajaKoejaksoResource(
                 )
         }
         return ResponseUtil.wrapOrNotFound(loppukeskusteluDTO)
-    }
-
-    @PutMapping("/koejakso/loppukeskustelu")
-    fun updateLoppukeskustelu(
-        @Valid @RequestBody loppukeskusteluDTO: KoejaksonLoppukeskusteluDTO,
-        principal: Principal?
-    ): ResponseEntity<KoejaksonLoppukeskusteluDTO> {
-        validateId(loppukeskusteluDTO.id, ENTITY_KOEJAKSON_LOPPUKESKUSTELU)
-        val user = userService.getAuthenticatedUser(principal)
-
-        var loppukeskustelu =
-            koejaksonLoppukeskusteluService.findOneByIdAndLahikouluttajaUserId(
-                loppukeskusteluDTO.id!!,
-                user.id!!
-            )
-
-        if (!loppukeskustelu.isPresent) {
-            loppukeskustelu =
-                koejaksonLoppukeskusteluService.findOneByIdAndLahiesimiesUserId(
-                    loppukeskusteluDTO.id!!,
-                    user.id!!
-                )
-
-            if (!loppukeskustelu.isPresent) {
-                throw BadRequestAlertException(
-                    "Koejakson loppukeskustelua ei löydy.",
-                    ENTITY_KOEJAKSON_LOPPUKESKUSTELU,
-                    "dataillegal.koejakson-loppukeskustelua-ei-loydy"
-                )
-            }
-
-            if (loppukeskustelu.get().lahikouluttaja?.sopimusHyvaksytty != true) {
-                throw BadRequestAlertException(
-                    "Esimies ei saa muokata loppukeskustelua, " +
-                        "jos kouluttaja ei ole allekirjoittanut niitä",
-                    ENTITY_KOEJAKSON_LOPPUKESKUSTELU,
-                    "dataillegal.esimies-ei-saa-muokata-loppukeskustelua-jos-kouluttaja-ei-ole-allekirjoittanut-niita"
-                )
-            }
-        }
-
-        validateArviointi(
-            loppukeskustelu.get().lahiesimies?.sopimusHyvaksytty,
-            ENTITY_KOEJAKSON_LOPPUKESKUSTELU
-        )
-
-        val result = koejaksonLoppukeskusteluService.update(loppukeskusteluDTO, user.id!!)
-        return ResponseEntity.ok(result)
-    }
-
-    private fun validateArviointi(
-        hyvaksytty: Boolean?,
-        entity: String
-    ) {
-        if (hyvaksytty == true) {
-            throw BadRequestAlertException(
-                "Hyväksyttyä arviointia ei saa muokata",
-                entity,
-                "dataillegal.hyvaksyttya-arviointia-ei-saa-muokata"
-            )
-        }
-    }
-
-    private fun validateId(id: Long?, entity: String) {
-        if (id == null) {
-            throw BadRequestAlertException(
-                "Virheellinen id",
-                entity,
-                "idnull"
-            )
-        }
     }
 }
